Make `Closer#manage`'s `closeable` argument non-nullable
This is done because the underlying default implementations used by the
factory methods (`#create` & `#createConcurrent`) use `ArrayDeque` and
`ConcurrentLinkedDeque` respectively, which both do not allow `null`
elements.

This would cause an NPE to be thrown when added.

The check for a `null` result from `Deque#pollFirst` is also redundant,
due to the previous call to `Deque#isEmpty`.

Signed-off-by: Mechite <contact@mechite.com>