Initial commit for additions to `stellar.serialization`
- Add `stellar.serialization.msgpack` as fork of `msgpack-java`
  (original portions remain licensed under original license)

- Rename existing one to use `.jsonb` suffix, as the `avaje-jsonb`
  adapter for use of this library

- Entirely remove Git submodule for `msgpack-java`

- Portions of test suite missing or disabled to push as a succeeding
  build - this must later be amended to complete the beginning of the
  development of the fork.

This fork may not be maintained forever - it may be replaced with a
custom implementation based on MPack <ludocode/mpack> C library.

When/if this is created, it will be made under `.messagepack` module to
prevent clashing, and the `.msgpack` one will later be removed, or most
likely just deprecated.

Signed-off-by: Mahied Maruf (mechite) <contact@mechite.com>