Delete `msgpack-java` submodule sources from source tree
They were left in order to reference them to slowly and correctly port
the entire test suite. Porting the test suite is unnessecary work,
considering this library was not modified except for clear refactorings,
the original test suite is passing, and the plan is to create a
custom parser implementation anyway.

Signed-off-by: Mahied Maruf (mechite) <contact@mechite.com>