Refactor `stellar.esthree.server` -> `stellar.esthree.mock`
The previous name will later be used for an actual server implementation,
whereas the mock server is to be used as a facet to test against, to
ensure the quality of our client implementation.

For testing the custom server implementation, we will simply use the
existing AWS SDK v2 to test against.

Signed-off-by: Mahied Maruf <contact@mechite.com>