Add working POC `stellar.lifecycle.spi`
It is recommended this module is not used. Favour avaje-spi-service.

It is added to the version history to be preserved and to serve as a
good proof-of-concept that I personally find cleaner than the source
code of avaje-spi-service.

- avaje-spi-service supports merging with manually written
  `META-INF/services`, this module does not

- avaje-spi-service has a comprehensive test suite,
  this module does not

This module was written to support a concept using Maven classifiers,
which has been cancelled and related code will be removed too.

Signed-off-by: Mahied Maruf <contact@mechite.com>