Remove `stellar.logging.bundle` entirely (see details)
This module is offensive to JPMS. There are two options that could have
been taken other than deleting it, both undesirable:

- Making this into a packaging=jar module, bundling them all and running
  the transformer ourselves, and providing a seperated module for JPMS
  `stellar.logging.bundle` that includes the logger, all official
  dispatchers and collectors.

  - This negates the modular nature of this system for those users
  - It prevents them from using any other collectors that are not in
    this monorepo.

- Leaving it as-is, updating the information to be clearer to say that
  this was designed with uber-JAR or WAR packaging in mind, and in the
  case of an uber-JAR, said META-INF merging transformers need to be
  run in order to allow for dispatcher & collector registration.

The best solution is to remove this module entirely, and allow users to
add only the dependencies that they require. Furthermore, the creation
of a BOM in 86327b09159cdbf27b80ea8984e6f5badc66b13c allows for this to
be far less verbose (the only valid reason there is for using bundle).

Fixes: #10
Related-to: 86327b09159cdbf27b80ea8984e6f5badc66b13c
Signed-off-by: Mahied Maruf <contact@mechite.com>