Create `stellar.annotation` and `stellar.annotation.intellij`
The IntelliJ Platform plugin adds synthetic JetBrains annotations to the
PSI to provide the same static analysis functionality.

This commit also extracts/removes the related/previous annotations from
`stellar.facility`.

The plugin is built with Maven Assembly Plugin to form a ZIP file in the
format that IntelliJ requires. Gradle is not used at all in the process.

Signed-off-by: Mahied Maruf <contact@mechite.com>