Create `LoggerFormat` and remove `Strings#format` (see details)
Move `Strings#format` and the associated tests, to be a part of the
logging API, as this is an implementation detail of the logger.

The new utility is also a SPI (service provider interface) which allows
for it to be substituted / extended with another implementation, if
somebody is having issues with the existing one.

In places where it was being used previously, this has been updated to
use either `java.lang.String#format`, or `MessageFormat`, depending on
what is the correct case.

In logging-related places, the dual-purpose utility `LoggerFormat` is
available to unify the formatting.

Other unrelated changes:

- Remove the use of the `Orderable` interface in `LoggerMessage`, and
  directly use the Comparable#compareTo available on `Instant`.

- Remove `ImmutableModificationException` from `stellar.facility` as it
  is no longer used and `IllegalStateException` is orthogonal

- Update documentation in `sirius.stellar.facility.terminal` to no
  longer reference a particular string formatter implementation

Signed-off-by: Mahied Maruf <contact@mechite.com>