Migrate from `sirius.stellar` -> `org.siriushq.stellar` (see details)
Change of `groupId` to `org.siriushq.stellar`, while maintaining the
same Java package and module names for now.

This commit is the minimal change required to publish to Maven Central.

This commit also modifies the `readme` and `license`, to reference the
company name & to no longer tell users to reference a custom repository,
and drops the concept of having a separate `stellar` vs `stellar-bom`
artifact - the `-bom` one is no longer available.

Signed-off-by: Mahied Maruf <contact@mechite.com>