XXE hardening of `TransformerFactory` in `sirius.stellar.esthree`
While the `DocumentBuilderFactory` had implementation-agnostic XXE
hardening applied (hardening that is expected to be available on all
JVM platforms, as specified for javax.xml), this was not the case for
the `TransformerFactory`.

This is generally not insecure, as the thread-local transformers created
by this factory are exclusively used for writing outputs.

It can a valid issue if unescaped input is provided to the methods
exposed by the `Esthree` interface, and there was any lack of escaping
inside of `Document`, `DOMSource`, `StreamResult`, etc, where XSLT
content can be injected, etc, and a DTD retrieved (recursively).

Related-to: CVE-2014-0107
Related-to: CVE-2017-7503
Related-to: CVE-2017-7465
Related-to: CVE-2021-41042
Signed-off-by: Mahied Maruf <contact@mechite.com>