Update formatting and documentation in `stellar.configuration`
Small refactoring to variable names, code formatting, and documentation,
which one reason for, is to wrap better to a lower column editor.

There is no strict rule to wrap within 80 columns in this code base due
to limitations of the syntax of Java, but it is a strong recommendation.

The idiom `it` is used instead of `value` for the lambda code example
in `Configuration`, making the example appear even smaller.

Other than this, `SignalConfigurationReloader` no longer implements the
`Runnable` interface; the `#run` method is given a descriptive name and
documentation, and implemented using a method reference instead,
for a simpler relationship.

Signed-off-by: Mahied Maruf <contact@mechite.com>