Favour a static `.class` over `Object#getClass`
In `stellar.configuration`, for preceeding provider lookup (where a
configuration provider can have dependencies, declared statically by
referencing those by their `.class`es), previously, for obtaining
the actual provider, `#getClass` is always used.

Now, a `#clazz` method is added to the `ConfigurationProvider`
interface, and all default implementations implement it statically.

The default implementation of this method still uses `Object#getClass`,
but this is a semantic change where it makes more sense to also declare
the self class of an implementation statically, as the only purpose of
the lookup is for dependency resolution, not reflection.

This simply offers a more flexible extensibility system, e.g.:

A | File_MyConfigurationProvider                  <- (marked as the SPI)
B | Network_MyConfigurationProvider               <- (marked as the SPI)
C | MyConfigurationProvider            <- (marked as a dependency point)
D | ConfigurationProvider                            <- (base interface)

...where there are two SPIs, A & B based on the base class C, but other
people can declare a dependency on C and still always be placed before
both A & B, as they extend #clazz providing C.class, not A.class etc.

From a performance standpoint, it is very unlikely that both
methods (#clazz or #getClass) would have any measurable difference,
due to the optimization of the #getClass lookup, and initial overhead
(pre-JIT) of the virtual method invocation, in this inheritance tree.

Signed-off-by: Mahied Maruf <contact@mechite.com>